pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "papersearch"
version = "0.1.0"
description = "Paper aggregation system for RL/robotics research with MCP server"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Sam Schoedel"}
]

dependencies = [
    "arxiv>=2.1.0",
    "feedparser>=6.0.0",
    "pyzotero>=1.5.0",
    "anthropic>=0.30.0",
    "openai>=1.0.0",
    "thefuzz>=0.20.0",
    "python-Levenshtein>=0.21.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
